updated changes

+++ server/tests/db.helper.js
@@ -4,15 +4,23 @@ const mongoose = require('mongoose');
 let mongod;
 
 module.exports.connect = async () => {
-  mongod = await MongoMemoryServer.create();
-  const uri = mongod.getUri();
-  await mongoose.connect(uri);
+  if (mongoose.connection.readyState === 0) {
+    mongod = await MongoMemoryServer.create();
+    const uri = mongod.getUri();
+    await mongoose.connect(uri);
+  }
 };
 
 module.exports.closeDatabase = async () => {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
-  await mongod.stop();
+  if (mongoose.connection.readyState !== 0) {
+    try {
+      await mongoose.connection.dropDatabase();
+    } catch (e) {}
+    await mongoose.connection.close();
+  }
+  if (mongod) {
+    await mongod.stop();
+  }
 };
 
 module.exports.clearDatabase = async () => {
